requirements/common.txt
Django==1.11.3
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.3
drf-extensions==0.3.1
drf-nested-routers==0.90.0
Markdown==2.6.8
olefile==0.44
Pillow==4.2.1
pytz==2017.2
